+++ sixtyfps_runtime/corelib/layout.rs
@@ -518,7 +518,7 @@ pub fn solve_box_layout(data: &BoxLayoutData, repeater_indexes: Slice<u32>) -> S
 
     let mut result = SharedVector::<f32>::default();
     result.resize(data.cells.len() * 2 + repeater_indexes.len(), 0.);
-    let res = result.as_slice_mut();
+    let res = result.as_mut_slice();
 
     // The index/2 in result in which we should add the next repeated item
     let mut repeat_offset =
@@ -650,7 +650,7 @@ pub fn solve_path_layout(data: &PathLayoutData, repeater_indexes: Slice<u32>) ->
 
     let mut result = SharedVector::<f32>::default();
     result.resize(data.item_count as usize * 2 + repeater_indexes.len(), 0.);
-    let res = result.as_slice_mut();
+    let res = result.as_mut_slice();
 
     // The index/2 in result in which we should add the next repeated item
     let mut repeat_offset =